specification: [refactor] Move specification types to @types
