+++ src/@types/syntax/elementSpecification.d.ts
@@ -1,3 +1,6 @@
+import { TData } from './data';
+import { IElementData, IElementExpression, IElementStatement, IElementBlock } from './elementsCore';
+
 export type TElementCategoryData = 'value' | 'boxidentifier';
 export type TElementCategoryExpression = 'operator-math';
 export type TElementCategoryStatement = 'box';
@@ -40,3 +43,43 @@ export type TElementName =
     | TElementExpressionName
     | TElementStatementName
     | TElementBlockName;
+
+export interface IElementDataSpecification {
+    label: string;
+    type: 'Data';
+    category: TElementCategoryData;
+    prototype: (name: TElementDataName, label: string) => IElementData<TData>;
+}
+
+export interface IElementExpressionSpecification {
+    label: string;
+    type: 'Expression';
+    category: TElementCategoryExpression;
+    prototype: (name: TElementExpressionName, label: string) => IElementExpression<TData>;
+}
+
+interface IElementInstructionSpecification {
+    allowAbove?: TElementName[] | boolean;
+    allowBelow?: TElementName[] | boolean;
+    forbidAbove?: TElementName[] | boolean;
+    forbidBelow?: TElementName[] | boolean;
+    allowedNestLevel?: number[] | 'any';
+    allowedNestInside?: TElementBlockName[] | boolean;
+    forbiddenNestInside?: TElementBlockName[] | boolean;
+}
+
+export type IElementStatementSpecification = IElementInstructionSpecification & {
+    label: string;
+    type: 'Statement';
+    category: TElementCategoryStatement;
+    prototype: (name: TElementStatementName, label: string) => IElementStatement;
+};
+
+export type IElementBlockSpecification = IElementInstructionSpecification & {
+    label: string;
+    type: 'Block';
+    category: TElementCategoryBlock;
+    prototype: (name: TElementBlockName, label: string) => IElementBlock;
+    allowNestInside?: TElementBlockName[] | boolean;
+    forbidNestInside?: TElementBlockName[] | boolean;
+};
